// src/generated/models/UserRankDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Department } from './Department';
export type UserRankDto = {
    rank?: number;
    userId?: string;
    username?: string;
    profileImage?: string;
    department?: Department;
    points?: number;
};